+++ crates/emmylua_ls/src/handlers/test/definition_test.rs
@@ -1,5 +1,7 @@
 #[cfg(test)]
 mod tests {
+    use lsp_types::GotoDefinitionResponse;
+
     use crate::handlers::test_lib::ProviderVirtualWorkspace;
 
     #[test]
@@ -71,5 +73,69 @@ mod tests {
                 print(t.abc<??>)
             "#,
         );
+    }
+
+    #[test]
+    fn test_goto_overload() {
+        let mut ws = ProviderVirtualWorkspace::new();
+        ws.def(
+            r#"
+                ---@class Goto1
+                ---@class Goto2
+                ---@class Goto3
+
+                ---@class T
+                ---@field func fun(a:Goto1) # 1
+                ---@field func fun(a:Goto2) # 2
+                ---@field func fun(a:Goto3) # 3
+                local T = {}
+
+                function T:func(a)
+                end
+            "#,
+        );
+
+        {
+            let result = ws
+                .check_definition(
+                    r#"
+                ---@type Goto2
+                local Goto2
+
+                ---@type T
+                local t
+                t.fu<??>nc(Goto2)
+                 "#,
+                )
+                .unwrap();
+            match result {
+                GotoDefinitionResponse::Array(array) => {
+                    assert_eq!(array.len(), 2);
+                }
+                _ => {
+                    panic!("expect array");
+                }
+            }
+        }
+
+        {
+            let result = ws
+                .check_definition(
+                    r#"
+                ---@type T
+                local t
+                t.fu<??>nc()
+                 "#,
+                )
+                .unwrap();
+            match result {
+                GotoDefinitionResponse::Array(array) => {
+                    assert_eq!(array.len(), 4);
+                }
+                _ => {
+                    panic!("expect array");
+                }
+            }
+        }
     }
 }